separate out ADR tests in BuildKite pipeline
