+++ buildkite/test-adr.dockerfile
@@ -7,5 +7,4 @@ ARG CODECOV_TOKEN
 ENV CODECOV_TOKEN=$CODECOV_TOKEN
 
 # Test app
-CMD ./gradlew :app:detektMain -PexcludeADR=true :app:test :userCLI:test :app:jacocoTestReport && \
-    codecov -f app/coverage/test/*.xml
+CMD ./gradlew :app:test --tests **ADRTests*
