+++ src/app/static/src/tests/integration/adr.itest.ts
@@ -1,8 +1,11 @@
-import {actions} from "../../app/store/baseline/actions";
+import {actions as baselineActions} from "../../app/store/baseline/actions";
+import {actions as surveyAndProgramActions} from "../../app/store/surveyAndProgram/actions";
 import {login, rootState} from "./integrationTest";
 import {BaselineMutation} from "../../app/store/baseline/mutations";
 import {actions as rootActions} from "../../app/store/root/actions";
 import {RootMutation} from "../../app/store/root/mutations";
+import {SurveyAndProgramMutation} from "../../app/store/surveyAndProgram/mutations";
+import {getFormData} from "./helpers";
 
 // this suite tests all endpoints that talk to the ADR
 // we put them in a suite of their own so that we can run
@@ -14,7 +17,12 @@ describe("ADR related actions", () => {
         await login();
         // this key is for a test user who has access to 1 fake dataset
         const adrKey = "4c69b103-4532-4b30-8a37-27a15e56c0bb"
-        await rootActions.saveADRKey({commit: jest.fn(), rootState} as any, adrKey);
+        await rootActions.saveADRKey({commit: jest.fn(), rootState} as any, adrKey)
+
+        const commit = jest.fn();
+        const dispatch = jest.fn();
+        const formData = getFormData("malawi.geojson");
+        await baselineActions.uploadShape({commit, dispatch, rootState} as any, formData);
     });
 
     it("can fetch ADR datasets", async () => {
@@ -22,7 +30,7 @@ describe("ADR related actions", () => {
         await rootActions.getADRDatasets({commit, rootState} as any);
 
         expect(commit.mock.calls[0][0]["type"]).toBe(RootMutation.SetADRDatasets);
-        expect(commit.mock.calls[0][0]["payload"]).toEqual([]);
+        expect(Array.isArray(commit.mock.calls[0][0]["payload"])).toBe(true);
     });
 
     it("can get dataset", async () => {
@@ -42,7 +50,7 @@ describe("ADR related actions", () => {
         await rootActions.getADRDatasets({commit, rootState} as any);
         const datasetId = commit.mock.calls[0][0]["payload"][0].id;
         const state = {selectedDataset: {id: datasetId}};
-        await actions.refreshDatasetMetadata({commit, state, dispatch, rootState: rootStateWithSchemas} as any);
+        await baselineActions.refreshDatasetMetadata({commit, state, dispatch, rootState: rootStateWithSchemas} as any);
         expect(commit.mock.calls[1][0]).toBe(BaselineMutation.UpdateDatasetResources);
     });
 
@@ -50,7 +58,7 @@ describe("ADR related actions", () => {
         const commit = jest.fn();
         const dispatch = jest.fn();
         const state = {country: "Malawi"} as any;
-        await actions.importPJNZ({commit, state, dispatch, rootState} as any,
+        await baselineActions.importPJNZ({commit, state, dispatch, rootState} as any,
             "https://raw.githubusercontent.com/mrc-ide/hint/master/src/app/testdata/Botswana2018.PJNZ");
 
         expect(commit.mock.calls[1][0]["type"]).toBe(BaselineMutation.PJNZUpdated);
@@ -61,7 +69,7 @@ describe("ADR related actions", () => {
     it("can import shape file", async () => {
         const commit = jest.fn();
         const dispatch = jest.fn();
-        await actions.importShape({commit, dispatch, rootState} as any,
+        await baselineActions.importShape({commit, dispatch, rootState} as any,
             "https://raw.githubusercontent.com/mrc-ide/hint/master/src/app/testdata/malawi.geojson");
 
         expect(commit.mock.calls[1][0]["type"]).toBe(BaselineMutation.ShapeUpdated);
@@ -73,12 +81,50 @@ describe("ADR related actions", () => {
     it("can import population file", async () => {
         const commit = jest.fn();
         const dispatch = jest.fn();
-        await actions.importPopulation({commit, dispatch, rootState} as any,
+        await baselineActions.importPopulation({commit, dispatch, rootState} as any,
             "https://raw.githubusercontent.com/mrc-ide/hint/master/src/app/testdata/population.csv");
 
         expect(commit.mock.calls[1][0]["type"]).toBe(BaselineMutation.PopulationUpdated);
         expect(commit.mock.calls[1][0]["payload"]["filename"])
             .toBe("population.csv");
+    }, 7000);
+
+    it("can import survey", async () => {
+
+        const commit = jest.fn();
+        const dispatch = jest.fn();
+
+        await surveyAndProgramActions.importSurvey({commit, dispatch, rootState} as any,
+            "https://raw.githubusercontent.com/mrc-ide/hint/master/src/app/testdata/survey.csv");
+
+        expect(commit.mock.calls[1][0]["type"]).toBe(SurveyAndProgramMutation.SurveyUpdated);
+        expect(commit.mock.calls[1][0]["payload"]["filename"])
+            .toBe("survey.csv")
+    }, 7000);
+
+    it("can import programme", async () => {
+
+        const commit = jest.fn();
+        const dispatch = jest.fn();
+
+        await surveyAndProgramActions.importProgram({commit, dispatch, rootState} as any,
+            "https://raw.githubusercontent.com/mrc-ide/hint/master/src/app/testdata/programme.csv");
+
+        expect(commit.mock.calls[1][0]["type"]).toBe(SurveyAndProgramMutation.ProgramUpdated);
+        expect(commit.mock.calls[1][0]["payload"]["filename"])
+            .toBe("programme.csv")
+    }, 7000);
+
+    it("can import anc", async () => {
+
+        const commit = jest.fn();
+
+        await surveyAndProgramActions.importANC({commit, rootState} as any,
+            "https://raw.githubusercontent.com/mrc-ide/hint/master/src/app/testdata/anc.csv");
+
+        expect(commit.mock.calls[1][0]["type"]).toBe(SurveyAndProgramMutation.ANCUpdated);
+        expect(commit.mock.calls[1][0]["payload"]["filename"])
+            .toBe("anc.csv");
     }, 7000);
 
 });
